manual effort way
