+++ src/scripts/components/dropbox/auth.html
@@ -1,0 +1,7 @@
+<div class="dropbox--auth">
+    <title>Login with Dropbox</title>
+    
+    <body>
+        test worked!
+    </body>
+</div>
