--- src/scripts/components/dropbox/settings/template.html
@@ -1,13 +1,0 @@
-<div class="form-group">
-    <h3>{{_.i18n('Dropbox API key')}}</h3>
-    <input class="form-control" type="text" name="dropboxKey" value="{{dropboxKey}}" placeholder="{{_.i18n(placeholder)}}" />
-</div>
-
-<div class="form-group">
-    <p>{{_.i18n('dropbox.api info 1')}} <a target="_blank" href="https://www.dropbox.com/developers/apps">Dropbox</a></p>
-    <p>{{_.i18n('dropbox.api info 2')}}</p>
-    <ol>
-        <li>{{_.i18n('dropbox.api info li 1')}}</li>
-        <li>{{_.i18n('dropbox.api info li 2')}}</li>
-    </ol>
-</div>
