+++ src/scripts/components/settings/show/sync/dropbox/template.html
@@ -1,0 +1,10 @@
+<div class="form-group">
+    <h3>Login with Dropbox</h3>
+    <p>To connect Encryptic with Dropbox please follow the steps below.</p>
+    <p>Open this URL in your browser to login to Dropbox:</p>
+    <p>{{authUrl}}</p>
+    <p>After, enter the code provided by Dropbox:</p>
+    <input class="form-control" id="dropboxAccessToken" type="text" name="dropboxAccessToken" value="{{dropboxAccessToken}}" placeholder="long code goes here" />
+</div>
+
+
